Adding shopping item CRUD integration tests Might need to refactor them a little because DRY

+++ monmonmvc/src/main/java/lt/galdebar/monmonmvc/service/ShoppingItemService.java
@@ -6,6 +6,7 @@ import lt.galdebar.monmonmvc.persistence.domain.dto.ShoppingCategoryDTO;
 import lt.galdebar.monmonmvc.persistence.domain.dto.ShoppingItemDTO;
 import lt.galdebar.monmonmvc.persistence.domain.dto.ShoppingKeywordDTO;
 import lt.galdebar.monmonmvc.persistence.repositories.ShoppingItemRepo;
+import lt.galdebar.monmonmvc.service.exceptions.shoppingitem.ShoppingItemNotFound;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -28,10 +29,12 @@ public class ShoppingItemService {
     private ShoppingItemCategoryService shoppingItemCategoryService;
 
     public ShoppingItemDTO addItem(ShoppingItemDTO shoppingItemDTO) {
-        ShoppingCategoryDTO foundCategory = shoppingItemCategoryService.findCategoryByKeyword(
-                new ShoppingKeywordDTO("", shoppingItemDTO.getItemName())
-        );
-        shoppingItemDTO.setItemCategory(foundCategory.getCategoryName());
+        if (shoppingItemDTO.getItemCategory().trim().isEmpty()) {
+            ShoppingCategoryDTO foundCategory = shoppingItemCategoryService.findCategoryByKeyword(
+                    new ShoppingKeywordDTO("", shoppingItemDTO.getItemName())
+            );
+            shoppingItemDTO.setItemCategory(foundCategory.getCategoryName());
+        }
         ShoppingItemDAO returnedItem = shoppingItemRepo.insert(dtoToDao(shoppingItemDTO));
         return daoToDto(returnedItem);
     }
@@ -40,8 +43,9 @@ public class ShoppingItemService {
         return shoppingItemRepo.findById(id);
     }
 
-    public List<ShoppingItemDAO> getItemsByCategory(String requestedCategory) {
-        return shoppingItemRepo.findByItemCategory(requestedCategory);
+    public List<ShoppingItemDTO> getItemsByCategory(String requestedCategory) {
+        List<ShoppingItemDAO> foundItems = shoppingItemRepo.findByItemCategory(requestedCategory);
+        return daosToDtos(foundItems);
 //        return shoppingItemRepo.findByItemCategoryAndUsers(requestedCategory, getCurrentUserAndConnectedUsers());
     }
 
@@ -50,36 +54,64 @@ public class ShoppingItemService {
     }
 
 
-    public ShoppingItemDTO updateItem(ShoppingItemDTO shoppingItemDTO) {
+    public ShoppingItemDTO updateItem(ShoppingItemDTO shoppingItemDTO) throws ShoppingItemNotFound {
+        if(!validateShoppingItemDTO(shoppingItemDTO)){
+            throw new ShoppingItemNotFound();
+        }
+        if(!shoppingItemRepo.existsById(shoppingItemDTO.getId())){
+            throw new ShoppingItemNotFound();
+        }
         ShoppingItemDAO result = shoppingItemRepo.save(dtoToDao(shoppingItemDTO));
         return daoToDto(result);
     }
 
-    public List<ShoppingItemDTO> updateItems(List<ShoppingItemDTO> shoppingItemDTOS) {
+    public List<ShoppingItemDTO> updateItems(List<ShoppingItemDTO> shoppingItemDTOS) throws ShoppingItemNotFound {
+        for(ShoppingItemDTO item: shoppingItemDTOS){
+            if(!validateShoppingItemDTO(item)){
+                throw new ShoppingItemNotFound();
+            }
+            if(!shoppingItemRepo.existsById(item.getId())){
+                throw new ShoppingItemNotFound();
+            }
+        }
         List<ShoppingItemDAO> updatedItems = shoppingItemRepo.saveAll(dtosToDaos(shoppingItemDTOS));
         return daosToDtos(updatedItems);
     }
 
-    public void deleteItem(ShoppingItemDTO shoppingItemDTO) {
+    public void deleteItem(ShoppingItemDTO shoppingItemDTO) throws ShoppingItemNotFound {
+        if(!validateShoppingItemDTO(shoppingItemDTO)){
+            throw new ShoppingItemNotFound();
+        }
+        if(!shoppingItemRepo.existsById(shoppingItemDTO.getId())){
+            throw new ShoppingItemNotFound();
+        }
         shoppingItemRepo.delete(dtoToDao(shoppingItemDTO));
     }
 
-    public void deleteItems(List<ShoppingItemDTO> shoppingItemDTOList) {
+    public void deleteItems(List<ShoppingItemDTO> shoppingItemDTOList) throws ShoppingItemNotFound {
+        for(ShoppingItemDTO item: shoppingItemDTOList){
+            if(!validateShoppingItemDTO(item)){
+                throw new ShoppingItemNotFound();
+            }
+            if(!shoppingItemRepo.existsById(item.getId())){
+                throw new ShoppingItemNotFound();
+            }
+        }
         shoppingItemRepo.deleteAll(dtosToDaos(shoppingItemDTOList));
     }
 
     private ShoppingItemDAO dtoToDao(ShoppingItemDTO shoppingItemDTO) {
         ShoppingItemDAO shoppingItemDAO = new ShoppingItemDAO();
         shoppingItemDAO.id = shoppingItemDTO.getId();
-                shoppingItemDAO.itemName = shoppingItemDTO.getItemName();
-                shoppingItemDAO.itemCategory = shoppingItemDTO.getItemCategory();
-                shoppingItemDAO.quantity = shoppingItemDTO.getQuantity();
-                shoppingItemDAO.comment = shoppingItemDTO.getComment();
-                shoppingItemDAO.isInCart = shoppingItemDTO.isInCart();
-                shoppingItemDAO.users.add(
-                        SecurityContextHolder.getContext().getAuthentication().getName()
-                );
-        return  shoppingItemDAO;
+        shoppingItemDAO.itemName = shoppingItemDTO.getItemName();
+        shoppingItemDAO.itemCategory = shoppingItemDTO.getItemCategory();
+        shoppingItemDAO.quantity = shoppingItemDTO.getQuantity();
+        shoppingItemDAO.comment = shoppingItemDTO.getComment();
+        shoppingItemDAO.isInCart = shoppingItemDTO.isInCart();
+        shoppingItemDAO.users.add(
+                SecurityContextHolder.getContext().getAuthentication().getName()
+        );
+        return shoppingItemDAO;
     }
 
     private List<ShoppingItemDAO> dtosToDaos(List<ShoppingItemDTO> shoppingItemDTOList) {
@@ -105,10 +137,24 @@ public class ShoppingItemService {
         return shoppingItemDTOList;
     }
 
-    private List<String> getCurrentUserAndConnectedUsers(){
+    private List<String> getCurrentUserAndConnectedUsers() {
         List<String> users = userService.getLinkedUsers();
         users.add(SecurityContextHolder.getContext().getAuthentication().getName());
         System.out.println(users);
         return users;
+    }
+
+    private boolean validateShoppingItemDTO(ShoppingItemDTO shoppingItemDTO) {
+        if(shoppingItemDTO == null){
+            return false;
+        }
+        if(shoppingItemDTO.getId() == null){
+            return false;
+        }
+        if(shoppingItemDTO.getId().trim().isEmpty()){
+            return false;
+        }
+
+        return true;
     }
 }
